+++ src/app/pages/(docs)/virtual-scroll.page.ts
@@ -1,5 +1,5 @@
 import { Component, ChangeDetectionStrategy, signal } from '@angular/core';
-import { VirtualScrollDirective } from 'quartz';
+import { VirtualScrollDirective } from '@quartz-headless/core';
 import { DemoPageComponent } from '../../components/demo-page/demo-page.component';
 import { CodeBlockComponent } from '../../components/code-block/code-block.component';
 import { BASIC_SNIPPET, API_SNIPPET } from './virtual-scroll.snippets';